move digital command to proper folder
